Refactored setting frag & vm
