+++ app/src/main/java/com/chrsrck/quakemap/ui/Settings.kt
@@ -1,4 +1,4 @@
-package com.chrsrck.quakemap
+package com.chrsrck.quakemap.ui
 
 import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.chrsrck.quakemap.R
+import com.chrsrck.quakemap.viewmodel.SettingsViewModel
 
 
 class Settings : Fragment() {
