+++ app/src/main/res/layout/settings_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Settings">
+    tools:context=".ui.Settings">
 
     <TextView
         android:layout_width="match_parent"
